pyproject.toml
[project]
name = "tooket-ther"
version = "0.1.0"
description = "Tooket-ther — concert ticketing backend (CN230)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "psycopg[binary]>=3.2.0",
    "PyJWT>=2.9.0",
    "httpx>=0.28.0",
    "redis>=5.2.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.3.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["tooket_ther*"]
